requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.2
python-docx==0.8.11
soupsieve==2.4.1
